requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
psutil
python-multipart
requests
streamlit
argparse